`tools/importer-rest-api-specs`: refactoring the existing pipeline to use the new `terraform` package
